Make simulator continue simulating after branch
Previously the simulator would just completely stop simulating after a
branching opcode was reached. Now, it marks the current stack (context
or function) as suppressed and only ignores simulation for that stack.
This means that if you have functions inside a function with branching,
the stacks for the inner functions will be simulated.